requirements.txt
# Зависимости бота Алиса
# Установка: pip install -r requirements.txt

# Discord
discord.py>=2.3.0          # Основная библиотека Discord
PyNaCl>=1.5.0              # ОБЯЗАТЕЛЬНО для голоса в Discord!

# Загрузка .env файла
python-dotenv>=1.0.0

# TTS — синтез речи (БЕСПЛАТНО)
edge-tts>=6.1.9            # Microsoft Edge TTS — основной
gTTS>=2.5.0                # Google TTS — запасной

# HTTP запросы (для Yandex API)
aiohttp>=3.9.0

# Музыка — скачивает аудио с SoundCloud, YouTube и др.
yt-dlp>=2024.1.0

# Генерация картинок — aiohttp уже выше (Pollinations AI не требует доп. библиотек)

# Погода — wttr.in (не требует доп. библиотек, только aiohttp)
